Update syncer tests
